+++ lib/weighted_list_rank/item.rb
@@ -11,5 +11,11 @@ module WeightedListRank
     def id
       raise NotImplementedError, "Implement this method to return the item's unique identifier"
     end
+
+    # Returns the score penalty of the item. Can be overridden by the including class.
+    # @return [Float, NilClass] the score penalty of the item as a percentage (e.g., 0.20 for 20%) or nil if no penalty
+    def score_penalty
+      nil
+    end
   end
 end