+++ test/test_helper.rb
@@ -25,11 +25,12 @@ module WeightedListRank
 
   class MockItem
     include Item
-    attr_reader :id, :position
+    attr_reader :id, :position, :score_penalty
 
-    def initialize(id, position)
+    def initialize(id, position, score_penalty = nil)
       @id = id
       @position = position
+      @score_penalty = score_penalty
     end
   end
 end